--- src/k8s/pkg/k8sd/features/images.go
@@ -1,16 +1,0 @@
-package features
-
-const dnsImageRepository = "ghcr.io/canonical/coredns"
-const dnsImageTag = "1.11.1-ck4"
-
-const ciliumAgentImageRepository = "ghcr.io/canonical/cilium"
-const ciliumAgentImageTag = "1.15.2-ck1"
-
-const ciliumOperatorImageRepository = "ghcr.io/canonical/cilium-operator"
-const ciliumOperatorImageTag = "1.15.2-ck1"
-
-const storageImageRepository = "ghcr.io/canonical/rawfile-localpv"
-const storageImageTag = "0.8.0-ck5"
-
-const metricsServerImageRepository = "ghcr.io/canonical/metrics-server"
-const metricsServerImageTag = "0.7.0-ck0"
