+++ src/k8s/pkg/k8sd/features/localpv/chart.go
@@ -1,0 +1,22 @@
+package localpv
+
+import (
+	"path"
+
+	"github.com/canonical/k8s/pkg/client/helm"
+)
+
+var (
+	// chart is manifests for the built-in local storage feature, powered by Rawfile LocalPV CSI.
+	chart = helm.InstallableChart{
+		Name:         "ck-storage",
+		Namespace:    "kube-system",
+		ManifestPath: path.Join("charts", "rawfile-csi-0.8.0.tgz"),
+	}
+
+	// imageRepo is the image to use for Rawfile LocalPV CSI.
+	imageRepo = "ghcr.io/canonical/rawfile-localpv"
+
+	// imageTag is the image tag to use.
+	imageTag = "0.8.0-ck5"
+)
