+++ src/k8s/pkg/k8sd/features/localpv/localpv.go
@@ -1,4 +1,4 @@
-package features
+package localpv
 
 import (
 	"context"
@@ -8,7 +8,6 @@ import (
 	"github.com/canonical/k8s/pkg/snap"
 )
 
-// ApplyLocalStorage is used to configure the Local Storage feature on Canonical Kubernetes.
 // ApplyLocalStorage deploys the rawfile-localpv CSI driver on the cluster based on the given configuration, when cfg.Enabled is true.
 // ApplyLocalStorage removes the rawfile-localpv when cfg.Enabled is false.
 // ApplyLocalStorage returns an error if anything fails.
@@ -27,14 +26,14 @@ func ApplyLocalStorage(ctx context.Context, snap snap.Snap, cfg types.LocalStora
 		"controller": map[string]any{
 			"csiDriverArgs": []string{"--args", "rawfile", "csi-driver", "--disable-metrics"},
 			"image": map[string]any{
-				"repository": storageImageRepository,
-				"tag":        storageImageTag,
+				"repository": imageRepo,
+				"tag":        imageTag,
 			},
 		},
 		"node": map[string]any{
 			"image": map[string]any{
-				"repository": storageImageRepository,
-				"tag":        storageImageTag,
+				"repository": imageRepo,
+				"tag":        imageTag,
 			},
 			"storage": map[string]any{
 				"path": cfg.GetLocalPath(),
@@ -42,6 +41,6 @@ func ApplyLocalStorage(ctx context.Context, snap snap.Snap, cfg types.LocalStora
 		},
 	}
 
-	_, err := m.Apply(ctx, chartLocalStorage, helm.StatePresentOrDeleted(cfg.GetEnabled()), values)
+	_, err := m.Apply(ctx, chart, helm.StatePresentOrDeleted(cfg.GetEnabled()), values)
 	return err
 }
